Create Meteorological data analysis
This project analyzes and visualizes the meteorological data in the coastal areas of northern Italy. In the course of the experiment, we will first use the matplotlib library in Python to graph the data, and finally draw our conclusions with the support of graph analysis. Research question: "How does the ocean affect the climate of an area"